add validations

+++ src/main/java/com/example/medrese/Service/ArticleService.java
@@ -2,6 +2,7 @@ package com.example.medrese.Service;
 
 import com.example.medrese.DTO.Request.Create.CreateArticleDTO;
 import com.example.medrese.DTO.Request.Update.UpdateArticle;
+import com.example.medrese.DTO.Response.ArticleProjection;
 import com.example.medrese.DTO.Response.ArticleResponse;
 import com.example.medrese.Model.Article;
 import com.example.medrese.Model.ArticleCategory;
@@ -13,6 +14,7 @@ import lombok.AllArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.ObjectUtils;
 
 import java.util.List;
 
@@ -29,14 +31,23 @@ public class ArticleService {
     AuthorArticleRepository authorArticleRepository;
     FileService fileService;
 
-    public List<ArticleResponse> getAllArticles() {
-        return articleRepository.findAll().stream()
-                .map(articleMapper::toResponse)
-                .toList();
+    public List<ArticleProjection> getAllArticles() {
+        return articleRepository.findAllArticlesWithAuthorsAndCategories();
+
     }
 
-    public Article getArticleById(Integer id) {
-        return articleRepository.findById(id).orElseThrow(() -> new RuntimeException("author not found"));
+    public ArticleResponse getArticleById(Integer id) {
+        Article article = articleRepository.findById(id).orElseThrow(() -> new RuntimeException("author not found"));
+        List<Integer> authors = authorArticleRepository.findByArticleId(article.getId()).stream()
+                .map(AuthorArticle::getAuthorId)
+                .toList();
+        List<Integer> categories = articleCategoryRepository.findByArticleId(article.getId()).stream()
+                .map(ArticleCategory::getCategoryId)
+                .toList();
+        ArticleResponse articleResponse = articleMapper.toResponse(article);
+        articleResponse.setAuthors(authors);
+        articleResponse.setCategories(categories);
+        return articleResponse;
     }
 
 
@@ -47,7 +58,13 @@ public class ArticleService {
         }
 
         Article article = articleMapper.toEntity(createArticleDTO);
+
+        if (fileService.isBase64(createArticleDTO.getImage())) {
+            article.setImage(fileService.uploadFile(createArticleDTO.getImage()));
+        }
+
         article = articleRepository.save(article);
+
 
         for (int category : createArticleDTO.getCategories()) {
             if (categoryRepository.existsById(category)) {
@@ -71,6 +88,7 @@ public class ArticleService {
         return articleMapper.toResponse(article);
     }
 
+    @Transactional
     public Article updateArticle(Integer id, UpdateArticle articleDetails) {
         Article article = articleRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Article not found with id " + id));
@@ -79,45 +97,60 @@ public class ArticleService {
             throw new RuntimeException("article with same title already exists");
         }
 
+        if (fileService.isBase64(articleDetails.getImage())) {
+            fileService.deleteFile(article.getImage());
+            String image = fileService.uploadFile(articleDetails.getImage());
+            article.setImage(image);
+        }
+
+
         article.setPublishedAt(articleDetails.getPublishedAt());
         article.setTitle(articleDetails.getTitle());
         article.setContent(articleDetails.getContent());
         article = articleRepository.save(article);
 
-        articleCategoryRepository.deleteByArticleIdIn(articleDetails.getCategories().stream().toList());
-        for (Integer category : articleDetails.getCategories()) {
-            if (articleCategoryRepository.existsByCategoryId(category)) {
-                ArticleCategory articleCategory = ArticleCategory.builder()
-                        .categoryId(category)
-                        .articleId(article.getId())
-                        .build();
-                articleCategoryRepository.save(articleCategory);
+        if(!ObjectUtils.isEmpty(articleDetails.getCategories())){
+            articleCategoryRepository.deleteByArticleIdIn(articleDetails.getCategories().stream().toList());
+            for (Integer category : articleDetails.getCategories()) {
+                if (articleCategoryRepository.existsByCategoryId(category)) {
+                    ArticleCategory articleCategory = ArticleCategory.builder()
+                            .categoryId(category)
+                            .articleId(article.getId())
+                            .build();
+                    articleCategoryRepository.save(articleCategory);
+                }
             }
         }
 
-        authorArticleRepository.deleteByAuthorIdInAndArticleId(articleDetails.getAuthorIds().stream().toList(),article.getId());
-        for (Integer author : articleDetails.getAuthorIds()) {
-            if (authorArticleRepository.existsByAuthorId(author)) {
-                AuthorArticle authorArticle = AuthorArticle.builder()
-                        .articleId(article.getId())
-                        .authorId(author)
-                        .build();
-                authorArticleRepository.save(authorArticle);
-            }
-        }
+       if(!ObjectUtils.isEmpty(articleDetails.getAuthorIds())){
+           authorArticleRepository.deleteByAuthorIdInAndArticleId(articleDetails.getAuthorIds().stream().toList(), article.getId());
+           for (Integer author : articleDetails.getAuthorIds()) {
+               if (authorArticleRepository.existsByAuthorId(author)) {
+                   AuthorArticle authorArticle = AuthorArticle.builder()
+                           .articleId(article.getId())
+                           .authorId(author)
+                           .build();
+                   authorArticleRepository.save(authorArticle);
+               }
+           }
+
+       }
 
 
         return article;
     }
 
     public void deleteArticle(Integer id) {
-        if (articleRepository.existsById(id)) {
-            throw new RuntimeException("does not exist");
-        }
+
+        Article article = articleRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Article not found with id " + id));
 
         articleCategoryRepository.deleteByArticleId(id);
 
-        articleRepository.deleteById(id);
+        articleRepository.delete(article);
+
+        fileService.deleteFile(article.getImage());
+
     }
